test status imports var table
