+++ config/defaults/profile_all.mx
@@ -307,8 +307,7 @@ autocomplete set git { [{
             #"IncFiles": true,
             "AllowMultiple": true,
             "Dynamic": ({
-                #git status -s -> regexp 'm/^( |\?)' -> [:1]
-                git status -s -> regexp 'f/^ . ["]?(.*?)["]?$/' -> cast str
+                git status -s -> regexp 'f/^[\ \?]. ["]?(.*?)["]?$/' -> cast str
             })
         }],
         "mv": [{
